Map Display
The map now displays in a graphical format. I have included a random
wall and floor texture I pulled from the internet for testing purposes.
For testing purposes you can move the screen around with W,A,S,D and
zoom in/out with the mouse wheel.
Also did some work in preparation for the character selection screen.