Sprites are now more uniform, added missing boss textures, map shows first sprite from the enemies
The enemy sprites should now be pretty close in size. Some of them might
be slightly wider/thinner to prevent distortion. Also added the 3
missing boss textures. Now the map shows the first sprite in the enemy
list for that room (ex. the boss). Also the ranger party is represented
by the first person in the party.